+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Optional support for serde
 - New type `Kelvin` for temperature (melting/boiling point)
-- New type `KiloJoulePerMol` for heat of fusion/evaporation
+- New type `KiloJoulePerMole` for heat of fusion/evaporation
 - New type `Picometer` for atomic radius
 - New type `Percent` for isotope abundance
 - New type `GramPerCubicCentimeter` for density
@@ -24,8 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Element::atomic_radius()` return type changed to `Option<Picometer>`
 - `Element::melting_point()` return type changed to `Option<Kelvin>`
 - `Element::boiling_point()` return type changed to `Option<Kelvin>`
-- `Element::fusion_heat()` return type changed to `Option<KiloJoulePerMol>`
-- `Element::evaporation_heat()` return type changed to `Option<KiloJoulePerMol>`
+- `Element::fusion_heat()` return type changed to `Option<KiloJoulePerMole>`
+- `Element::evaporation_heat()` return type changed to `Option<KiloJoulePerMole>`
 - `Isotope::natural_abundance()` return type changed to `Option<Percent>`
 
 ## [0.7.0] - 2023-02-06